requirements.txt
einops>=0.6.1
numpy>=1.24.4
opencv-python==4.6.0.66
pillow>=9.5.0
scipy==1.11.1
timm>=0.9.2
tokenizers
torch>=2.0.1

torchaudio>=2.0.2
torchdata==0.6.1
torchmetrics>=1.0.1
torchvision>=0.15.2

tqdm>=4.65.0
transformers==4.35.2
urllib3<1.27,>=1.25.4
xformers>=0.0.20
streamlit-keyup==0.2.0
lpips
clean-fid
peft
dominate
diffusers==0.25.1
gradio==3.43.1

vision_aided_loss
